+++ CHANGES.rst
@@ -240,6 +240,10 @@ Smart Rollup node
   multiple L1 blocks. The results are compared with what the rollup node has
   stored on disk. (MR :gl:`!18160`)
 
+- Produce Opentelemetry traces with ``--profiling``, configurable with a field
+  ``"otel_profiling"`` in the configuration. (MR :gl:`!18274`)
+
+
 Smart Rollup WASM Debugger
 --------------------------